# Conflicts:
#	Vereinsmanager.Server.Core/Controllers/DataTransferObjects/ScoreDto.cs
#	Vereinsmanager.Server.Core/Controllers/DataTransferObjects/VoiceDto.cs
#	Vereinsmanager.Server.Core/Database/ScoreManagment/Score.cs
#	Vereinsmanager.Server.Core/Migrations/20260223102738_Added_OAuth_and_UniqeIndex.Designer.cs
#	Vereinsmanager.Server.Core/Migrations/20260305105614_Initial.Designer.cs
#	Vereinsmanager.Server.Core/Migrations/20260311105327_Initial.Designer.cs
#	Vereinsmanager.Server.Core/Migrations/ServerDatabaseContextModelSnapshot.cs
#	Vereinsmanager.Server.Core/Program.cs
#	Vereinsmanager.Server.Core/Services/ScoreManagement/ScoreService.cs
#	Vereinsmanager.Server.Core/Utils/ErrorUtils.cs
#	Vereinsmanager.Server.Core/Vereinsmanager.Server.Core.csproj

+++ Vereinsmanager.Server.Core/Controllers/DataTransferObjects/ScoreDto.cs
@@ -8,7 +8,7 @@ public class ScoreDto : MetaDataDto
     public string Title { get; init; }
     public string Composer { get; init; }
     public string? Link { get; init; }
-    public int? Duration { get; init; }
+    public double? Duration { get; init; }
 
     public int[]? Sheets { get; init; }
